perf: cleanup reshape instrumentation, i32 indices for gather
- Remove reshape profiling counters
- Use i32 instead of i64 for gather indices (matches backend IntElem)
- Add f32→backend float conversion in Value::from_data
- Add backend float→f32 conversion in to_vec_f32
- Tested f16 backend (precision issues, reverted to f32)
- swap_dims no-op optimization when dims are equal

+++ inference/src/ops/tensor_manip.rs
@@ -238,8 +238,8 @@ pub fn gather_proto(
         })
         .unwrap_or(0);
 
-    // Get indices as i64 vec for GPU tensor
-    let idx_vec: Vec<i64> = indices.to_vec_f32().iter().map(|&v| v as i64).collect();
+    // Get indices as i32 vec for GPU tensor
+    let idx_vec: Vec<i32> = indices.to_vec_f32().iter().map(|&v| v as i32).collect();
     let indices_shape = indices.shape();
     let num_indices = idx_vec.len();
